requirements.txt
django-threadlocals
django-compat
